Added e2e transfer

+++ backend/src/main/java/com/netcracker/edu/back/backend/repository/ProductRepository.java
@@ -3,6 +3,7 @@ package com.netcracker.edu.back.backend.repository;
 import com.netcracker.edu.back.backend.entity.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Long> {
+
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 
 }
